OAE-18
A new CSS rule has been added which matches the selector label.required. This is not ideal because now both the input field needs to have class="required"
(to be properly handled by the validation code) and a label with class="required". However no other expedient CSS change could be found due to complications
styling input elements themselves.

This has not been pushed into a wider reaching "standard" throughout OAE as that will require a more thorough investigation of the handling of required fields
thoughout and testing of the UI in those cases to ensure satisfactory rendering. A follow-on issue will be created for that work.